fix a potential bug. string::npos is -1, size_t is 64 bits, but uint32_t is 32 bits. so in 64bit machine, 32 bit -1 is not equal 64 bit -1

+++ src/walt/walt.cpp
@@ -260,7 +260,7 @@ int main(int argc, const char **argv) {
       }
     }
 
-    uint32_t suffix_pos = output_file.find_last_of(".");
+    size_t suffix_pos = output_file.find_last_of(".");
     if (suffix_pos == string::npos) {
       SAM = true;
     } else if (".mr" == output_file.substr(suffix_pos)) {
